requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pymysql==1.1.1
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
bcrypt==4.2.1
python-multipart==0.0.20
jinja2==3.1.6
websockets==15.0.1
httpx==0.28.1
